Extended to include more path and extra decision. Still need to have it based on external rule file for final decision, currently it is using contraint fields.

+++ src/main/java/org/jbpm/demo/Request.java
@@ -4,7 +4,7 @@ public class Request {
 	
 	private String id;
 	private String personId;
-	private Long amount;
+	private Integer amount;
 	private boolean valid = true;
 	private String invalidReason;
 	private boolean canceled = false;
@@ -25,11 +25,11 @@ public class Request {
 		this.personId = personId;
 	}
 	
-	public Long getAmount() {
+	public Integer getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Long amount) {
+	public void setAmount(Integer amount) {
 		this.amount = amount;
 	}
